claude-agent-framework/python/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "claude-agent-framework"
version = "1.0.0"
description = "A comprehensive framework for building Claude-powered agents"
authors = [{name = "Anthropic", email = "support@anthropic.com"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "anthropic>=0.54.0",
    "aiohttp>=3.8.0",
    "aiostream>=0.4.5",
    "jsonschema>=4.22.0",
    "click>=8.0.0",
    "pyyaml>=6.0",
    "mcp>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[project.scripts]
claude-agent = "claude_agent.cli.main:main"

[project.urls]
Homepage = "https://github.com/anthropic/claude-agent-framework"
Documentation = "https://claude-agent-framework.readthedocs.io"
Repository = "https://github.com/anthropic/claude-agent-framework"
Issues = "https://github.com/anthropic/claude-agent-framework/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["claude_agent*"]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
